backend/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1
pydantic==2.9.2
python-dotenv==1.0.1
sqlparse==0.5.1
httpx==0.27.2
